renamed and moved the spotify and youtube api functions
